Split the Benchmark itself off
