+++ test/bench/util/CMakeLists.txt
@@ -2,4 +2,5 @@ add_library(util STATIC timer.cpp
                         results.cpp
                         random.cpp
                         benchmark.cpp
+                        realmbench.cpp
 )
